D3ASIM-1024: Final changes.
